+++ tests/test_integration/test_config.py
@@ -1,4 +1,3 @@
-import itertools
 import os
 import subprocess
 import textwrap
@@ -84,49 +83,8 @@ def test_config_not_used(repo):
         get_version_script(repo)
 
 
-@pytest.mark.parametrize(
-    "value",
-    [True, False],
-)
-def test_config_bool_pyproject_toml(repo, value):
-    create_pyproject_toml(repo, value)
-
-    with pytest.raises(subprocess.CalledProcessError):
-        get_version(repo)
-
-    with pytest.raises(subprocess.CalledProcessError):
-        get_version_module(repo)
-
-    with pytest.raises(subprocess.CalledProcessError):
-        get_version_script(repo)
-
-
-@pytest.mark.parametrize(
-    "option",
-    ["version_config", "setuptools_git_versioning"],
-)
-def test_config_false_setup_py(repo, option):
-    create_setup_py(repo, False, option=option)
-
-    assert get_version_setup_py(repo) == "0.0.0"
-
-
-@pytest.mark.parametrize(
-    "option",
-    ["version_config", "setuptools_git_versioning"],
-)
-def test_config_true_setup_py(repo, option):
-    create_setup_py(repo, True, option=option)
-
-    assert get_version_setup_py(repo) == "0.0.1"
-
-
-@pytest.mark.parametrize(
-    "option",
-    ["version_config", "setuptools_git_versioning"],
-)
-def test_config_enabled_true(repo, create_config, option):
-    create_config(repo, {"enabled": True}, option=option)
+def test_config_enabled_true(repo, create_config):
+    create_config(repo, {"enabled": True})
 
     assert get_version(repo) == "0.0.1"
     assert get_version_script(repo) == "0.0.1"
@@ -138,23 +96,29 @@ def test_config_enabled_true(repo, create_config, option):
 
 
 @pytest.mark.parametrize(
-    "option",
-    ["version_config", "setuptools_git_versioning"],
+    "value",
+    [
+        True,
+        False,
+        [("A", "B")],
+    ],
 )
-def test_config_wrong_format(repo, create_config, option):
-    create_config(repo, [("A", "B")], option=option)
+def test_config_wrong_format(repo, create_config, value):
+    create_config(repo, value)
 
     with pytest.raises(subprocess.CalledProcessError):
         get_version(repo)
 
+    with pytest.raises(subprocess.CalledProcessError):
+        get_version_module(repo)
 
-@pytest.mark.parametrize(
-    "option",
-    ["version_config", "setuptools_git_versioning"],
-)
-def test_config_pyproject_toml_is_used_then_setup_py_is_empty(repo, option):
+    with pytest.raises(subprocess.CalledProcessError):
+        get_version_script(repo)
+
+
+def test_config_pyproject_toml_is_used_then_setup_py_is_empty(repo):
     create_pyproject_toml(repo, {"starting_version": "2.3.4"}, add=False, commit=False)
-    create_setup_py(repo, NotImplemented, option=option, add=False, commit=False)
+    create_setup_py(repo, NotImplemented, add=False, commit=False)
 
     assert get_version(repo) == "2.3.4"
     assert get_version_script(repo) == "2.3.4"
@@ -233,63 +197,6 @@ def test_config_pyproject_toml_is_folder(repo):
 
     with pytest.raises(subprocess.CalledProcessError):
         get_version(repo)
-
-    with pytest.raises(subprocess.CalledProcessError):
-        get_version_module(repo)
-
-    with pytest.raises(subprocess.CalledProcessError):
-        get_version_script(repo)
-
-
-@pytest.mark.parametrize(
-    "version_config, setuptools_git_versioning",
-    itertools.combinations(
-        [
-            False,
-            True,
-            {"enabled": True},
-            {"enabled": False},
-            {"any": "abc"},
-        ],
-        2,
-    ),
-)
-def test_config_both_old_and_new_config_are_set(repo, version_config, setuptools_git_versioning):
-    create_file(
-        repo,
-        "setup.py",
-        textwrap.dedent(
-            """
-            from coverage.control import Coverage
-
-            coverage = Coverage()
-            coverage.start()
-
-            try:
-                import setuptools
-
-                setuptools.setup(
-                    name="mypkg",
-                    setuptools_git_versioning={setuptools_git_versioning},
-                    version_config={version_config},
-                    setup_requires=[
-                        "setuptools>=41",
-                        "wheel",
-                        "setuptools-git-versioning",
-                    ]
-                )
-            finally:
-                coverage.stop()
-                coverage.save()
-            """
-        ).format(version_config=version_config, setuptools_git_versioning=setuptools_git_versioning),
-    )
-
-    with pytest.raises(subprocess.CalledProcessError):
-        get_version(repo)
-
-    with pytest.raises(subprocess.CalledProcessError):
-        get_version_setup_py(repo)
 
     with pytest.raises(subprocess.CalledProcessError):
         get_version_module(repo)